REF: Move transaction class to own module.
