+++ tests/test_perf_tracking.py
@@ -36,7 +36,7 @@ from six.moves import range, zip
 import zipline.utils.factory as factory
 import zipline.finance.performance as perf
 from zipline.finance.performance import position_tracker
-from zipline.finance.slippage import Transaction, create_transaction
+from zipline.finance.transaction import Transaction, create_transaction
 import zipline.utils.math_utils as zp_math
 
 from zipline.gens.composites import date_sorted_sources
