+++ zipline/finance/slippage.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2014 Quantopian, Inc.
+# Copyright 2015 Quantopian, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ from functools import partial
 
 from six import with_metaclass
 
-from zipline.protocol import DATASOURCE_TYPE
+from zipline.finance.transaction import create_transaction
 from zipline.utils.serialization_utils import (
     VERSION_LABEL
 )
@@ -50,66 +50,6 @@ def transact_stub(slippage, commission, event, open_orders):
 
 def transact_partial(slippage, commission):
     return partial(transact_stub, slippage, commission)
-
-
-class Transaction(object):
-
-    def __init__(self, sid, amount, dt, price, order_id, commission=None):
-        self.sid = sid
-        self.amount = amount
-        self.dt = dt
-        self.price = price
-        self.order_id = order_id
-        self.commission = commission
-        self.type = DATASOURCE_TYPE.TRANSACTION
-
-    def __getitem__(self, name):
-        return self.__dict__[name]
-
-    def to_dict(self):
-        py = copy(self.__dict__)
-        del py['type']
-        return py
-
-    def __getstate__(self):
-
-        state_dict = copy(self.__dict__)
-
-        STATE_VERSION = 1
-        state_dict[VERSION_LABEL] = STATE_VERSION
-
-        return state_dict
-
-    def __setstate__(self, state):
-
-        OLDEST_SUPPORTED_STATE = 1
-        version = state.pop(VERSION_LABEL)
-
-        if version < OLDEST_SUPPORTED_STATE:
-            raise BaseException("Transaction saved state is too old.")
-
-        self.__dict__.update(state)
-
-
-def create_transaction(event, order, price, amount):
-
-    # floor the amount to protect against non-whole number orders
-    # TODO: Investigate whether we can add a robust check in blotter
-    # and/or tradesimulation, as well.
-    amount_magnitude = int(abs(amount))
-
-    if amount_magnitude < 1:
-        raise Exception("Transaction magnitude must be at least 1.")
-
-    transaction = Transaction(
-        sid=event.sid,
-        amount=int(amount),
-        dt=event.dt,
-        price=price,
-        order_id=order.id
-    )
-
-    return transaction
 
 
 class LiquidityExceeded(Exception):
